feat: Redis Pub/Sub 기반 알림 실시간 전달 구조로 전환
- NotificationEntity JPA 엔티티 추가 및 DB 영속화로 알림 유실 방지
- @TransactionalEventListener(AFTER_COMMIT) + @Async로 DB 커밋 이후 Redis PUBLISH 분리
- RedisMessageListenerContainer로 Pub/Sub 구독, DeferredResult 즉시 응답
- waitForUserNotifications() DB-first 조회로 Redis 장애 시 복원력 확보
- NotificationPubSubConfig 분리로 RedisConfig 순환의존 방지
- Post @Builder.Default 추가로 빌더 사용 시 기본값 보장
- PostViewCountBatchService long→int 캐스트 버그 수정

+++ src/main/java/com/economy/community/service/PostViewCountBatchService.java
@@ -48,7 +48,7 @@ public class PostViewCountBatchService {
             if (rawValue != null) {
                 long increment = Long.parseLong(rawValue);
                 if (increment > 0) {
-                    postRepository.updatePostViewCount(postId, increment);
+                    postRepository.updatePostViewCount(postId, (int) increment);
                 }
             }
         }
